requirements.txt
depthai==2.1.0.0
numpy==1.19.5
opencv-python==4.5.1.48
pyyaml==5.4.1
